+++ packages/eas-cli/src/worker/upload.ts
@@ -1,64 +1,32 @@
+import cliProgress from 'cli-progress';
 import * as https from 'https';
 import createHttpsProxyAgent from 'https-proxy-agent';
-import mime from 'mime';
-import { Gzip } from 'minizlib';
-import fetch, { Headers, HeadersInit, RequestInit, Response } from 'node-fetch';
-import fs, { createReadStream } from 'node:fs';
-import { readFile } from 'node:fs/promises';
-import path from 'node:path';
+import fetch, { BodyInit, Headers, HeadersInit, RequestInit, Response } from 'node-fetch';
+import fs from 'node:fs';
+import os from 'node:os';
+import { Readable } from 'node:stream';
 import promiseRetry from 'promise-retry';
 
+import { AssetFileEntry } from './assets';
+import { createMultipartBodyFromFilesAsync, multipartContentType } from './utils/multipart';
+
+const MAX_CONCURRENCY = Math.min(10, Math.max(os.availableParallelism() * 2, 20));
 const MAX_RETRIES = 4;
-const MAX_CONCURRENCY = 10;
-const MIN_RETRY_TIMEOUT = 100;
-const MAX_UPLOAD_SIZE = 5e8; // 5MB
-const MIN_COMPRESSION_SIZE = 5e4; // 50kB
 
-const isCompressible = (contentType: string | null, size: number): boolean => {
-  if (size < MIN_COMPRESSION_SIZE) {
-    // Don't compress small files
-    return false;
-  } else if (contentType && /^(?:audio|video|image)\//i.test(contentType)) {
-    // Never compress images, audio, or videos as they're presumably precompressed
-    return false;
-  } else if (contentType && /^application\//i.test(contentType)) {
-    // Only compress `application/` files if they're marked as XML/JSON/JS
-    return /(?:xml|json5?|javascript)$/i.test(contentType);
-  } else {
-    return true;
-  }
-};
+export type UploadPayload =
+  | { filePath: string }
+  | { asset: AssetFileEntry }
+  | { multipart: AssetFileEntry[] };
 
-const getContentTypeAsync = async (filePath: string): Promise<string | null> => {
-  let contentType = mime.getType(path.basename(filePath));
-
-  if (!contentType) {
-    const fileContent = await readFile(filePath, 'utf-8');
-    try {
-      // check if file is valid JSON without an extension, e.g. for the apple app site association file
-      const parsedData = JSON.parse(fileContent);
-
-      if (parsedData) {
-        contentType = 'application/json';
-      }
-    } catch {}
-  }
-
-  return contentType;
-};
-
-export interface UploadParams extends Omit<RequestInit, 'signal' | 'body'> {
-  filePath: string;
-  compress?: boolean;
-  url: string;
+export interface UploadRequestInit {
+  baseURL: string | URL;
   method?: string;
   headers?: HeadersInit;
-  body?: undefined;
   signal?: AbortSignal;
 }
 
 export interface UploadResult {
-  params: UploadParams;
+  payload: UploadPayload;
   response: Response;
 }
 
@@ -79,51 +47,67 @@ const getAgent = (): https.Agent => {
   }
 };
 
-export async function uploadAsync(params: UploadParams): Promise<UploadResult> {
-  const {
-    filePath,
-    signal,
-    compress,
-    method = 'POST',
-    url,
-    headers: headersInit,
-    ...requestInit
-  } = params;
-  const stat = await fs.promises.stat(filePath);
-  if (stat.size > MAX_UPLOAD_SIZE) {
-    throw new Error(
-      `Upload of "${filePath}" aborted: File size is greater than the upload limit (>500MB)`
-    );
-  }
+type OnProgressUpdateCallback = (progress: number) => void;
 
-  const contentType = await getContentTypeAsync(filePath);
-
+export async function uploadAsync(
+  init: UploadRequestInit,
+  payload: UploadPayload,
+  onProgressUpdate?: OnProgressUpdateCallback
+): Promise<UploadResult> {
   return await promiseRetry(
     async retry => {
-      const headers = new Headers(headersInit);
-      if (contentType) {
-        headers.set('content-type', contentType);
+      if (onProgressUpdate) {
+        onProgressUpdate(0);
       }
 
-      let bodyStream: NodeJS.ReadableStream = createReadStream(filePath);
-      if (compress && isCompressible(contentType, stat.size)) {
-        const gzip = new Gzip({ portable: true });
-        bodyStream.on('error', error => gzip.emit('error', error));
-        // @ts-expect-error: Gzip implements a Readable-like interface
-        bodyStream = bodyStream.pipe(gzip) as NodeJS.ReadableStream;
-        headers.set('content-encoding', 'gzip');
+      const headers = new Headers(init.headers);
+
+      const url = new URL(`${init.baseURL}`);
+      let errorPrefix: string;
+      let body: BodyInit | undefined;
+      let method = init.method || 'POST';
+      if ('asset' in payload) {
+        const { asset } = payload;
+        errorPrefix = `Upload of "${asset.normalizedPath}" failed`;
+        if (asset.type) {
+          headers.set('content-type', asset.type);
+        }
+        if (asset.size) {
+          headers.set('content-length', `${asset.size}`);
+        }
+        method = 'POST';
+        url.pathname = `/asset/${asset.sha512}`;
+        body = fs.createReadStream(asset.path);
+      } else if ('filePath' in payload) {
+        const { filePath } = payload;
+        errorPrefix = 'Worker deployment failed';
+        body = fs.createReadStream(filePath);
+      } else if ('multipart' in payload) {
+        const { multipart } = payload;
+        errorPrefix = `Upload of ${multipart.length} assets failed`;
+        headers.set('content-type', multipartContentType);
+        method = 'PATCH';
+        url.pathname = '/asset/batch';
+        const iterator = createMultipartBodyFromFilesAsync(
+          multipart.map(asset => ({
+            name: asset.sha512,
+            filePath: asset.path,
+            contentType: asset.type,
+            contentLength: asset.size,
+          })),
+          onProgressUpdate
+        );
+        body = Readable.from(iterator);
       }
 
       let response: Response;
       try {
-        response = await fetch(params.url, {
-          ...requestInit,
+        response = await fetch(url, {
           method,
-          body: bodyStream,
+          body,
           headers,
           agent: getAgent(),
-          // @ts-expect-error: Internal types don't match
-          signal,
+          signal: init.signal as any,
         });
       } catch (error) {
         return retry(error);
@@ -141,10 +125,10 @@ export async function uploadAsync(params: UploadParams): Promise<UploadResult> {
           if (rayId) {
             message += `\nReport this error quoting Request ID ${rayId}`;
           }
-          return `Upload of "${filePath}" failed: ${message}`;
+          return `${errorPrefix}: ${message}`;
         } else {
           const json = await response.json().catch(() => null);
-          return json?.error ?? `Upload of "${filePath}" failed: ${response.statusText}`;
+          return json?.error ?? `${errorPrefix}: ${response.statusText}`;
         }
       };
 
@@ -156,22 +140,23 @@ export async function uploadAsync(params: UploadParams): Promise<UploadResult> {
       ) {
         return retry(new Error(await getErrorMessageAsync()));
       } else if (response.status === 413) {
-        const message = `Upload of "${filePath}" failed: File size exceeded the upload limit`;
+        const message = `${errorPrefix!}: File size exceeded the upload limit`;
         throw new Error(message);
       } else if (!response.ok) {
         throw new Error(await getErrorMessageAsync());
+      } else if (onProgressUpdate) {
+        onProgressUpdate(1);
       }
 
       return {
-        params,
+        payload,
         response,
       };
     },
     {
       retries: MAX_RETRIES,
-      minTimeout: MIN_RETRY_TIMEOUT,
-      randomize: true,
-      factor: 2,
+      minTimeout: 50,
+      randomize: false,
     }
   );
 }
@@ -199,30 +184,53 @@ export async function callUploadApiAsync(url: string | URL, init?: RequestInit):
 }
 
 export interface UploadPending {
-  params: UploadParams;
+  payload: UploadPayload;
+  progress: number;
 }
 
-export type BatchUploadSignal = UploadResult | UploadPending;
-
 export async function* batchUploadAsync(
-  uploads: readonly UploadParams[]
-): AsyncGenerator<BatchUploadSignal> {
+  init: UploadRequestInit,
+  payloads: UploadPayload[],
+  onProgressUpdate?: OnProgressUpdateCallback
+): AsyncGenerator<UploadPending> {
+  const progressTracker = new Array(payloads.length).fill(0);
   const controller = new AbortController();
   const queue = new Set<Promise<UploadResult>>();
+  const initWithSignal = { ...init, signal: controller.signal };
+  const getProgressValue = (): number => {
+    const progress = progressTracker.reduce((acc, value) => acc + value, 0);
+    return progress / payloads.length;
+  };
+  const sendProgressUpdate =
+    onProgressUpdate &&
+    (() => {
+      onProgressUpdate(getProgressValue());
+    });
   try {
     let index = 0;
-    while (index < uploads.length || queue.size > 0) {
-      while (queue.size < MAX_CONCURRENCY && index < uploads.length) {
-        const uploadParams = uploads[index++];
-        let uploadPromise: Promise<UploadResult>;
-        queue.add(
-          (uploadPromise = uploadAsync({ ...uploadParams, signal: controller.signal }).finally(() =>
-            queue.delete(uploadPromise)
-          ))
+    while (index < payloads.length || queue.size > 0) {
+      while (queue.size < MAX_CONCURRENCY && index < payloads.length) {
+        const currentIndex = index++;
+        const payload = payloads[currentIndex];
+        const onChildProgressUpdate =
+          sendProgressUpdate &&
+          ((progress: number) => {
+            progressTracker[currentIndex] = progress;
+            sendProgressUpdate();
+          });
+        const uploadPromise = uploadAsync(initWithSignal, payload, onChildProgressUpdate).finally(
+          () => {
+            queue.delete(uploadPromise);
+            progressTracker[currentIndex] = 1;
+          }
         );
-        yield { params: uploadParams };
+        queue.add(uploadPromise);
+        yield { payload, progress: getProgressValue() };
       }
-      yield await Promise.race(queue);
+      yield {
+        ...(await Promise.race(queue)),
+        progress: getProgressValue(),
+      };
     }
 
     if (queue.size > 0) {
@@ -233,4 +241,25 @@ export async function* batchUploadAsync(
       throw error;
     }
   }
+}
+
+interface UploadProgressBar {
+  update(progress: number): void;
+  stop(): void;
+}
+
+export function createProgressBar(label = 'Uploading assets'): UploadProgressBar {
+  const queueProgressBar = new cliProgress.SingleBar(
+    { format: `|{bar}| {percentage}% ${label}` },
+    cliProgress.Presets.rect
+  );
+  queueProgressBar.start(1, 0);
+  return {
+    update(progress: number) {
+      queueProgressBar.update(progress);
+    },
+    stop() {
+      queueProgressBar.stop();
+    },
+  };
 }